+++ test/testUcumLhcUtils.spec.js
@@ -285,13 +285,13 @@ it("should return a message for invalid unit strings", function() {
     assert.equal(resp5.toVal.toPrecision(3), 2);
   });
 
-  it('should covert 2.[degF] to Cel', ()=>{
+  it('should convert 2.[degF] to Cel', ()=>{
     var resp = utils.convertUnitTo('2.[degF]', 1, 'Cel');
     assert.equal(resp.status, 'succeeded', resp.status + resp.msg);
     assert.equal(resp.toVal.toPrecision(7), -16.66667);
   });
 
-  it('should covert [degF].2 to Cel', ()=>{
+  it('should convert [degF].2 to Cel', ()=>{
     var resp = utils.convertUnitTo('[degF].2', 1, 'Cel');
     assert.equal(resp.status, 'succeeded', resp.status + resp.msg);
     assert.equal(resp.toVal.toPrecision(7), -16.66667);
@@ -337,7 +337,7 @@ describe('Test getSynonyms method', function() {
 }); // end checkSynonyms tests
 
 
-describe('toBaseUnits', ()=> {
+describe('convertToBaseUnits', ()=> {
   // Base units currently appear in the dimension vector in this order
   // (taken from UnitTables.getInstance().dimVecIndexToBaseUnit_)
   /*   '0': 'm',
@@ -348,71 +348,88 @@ describe('toBaseUnits', ()=> {
        '5': 'C',
        '6': 'cd' */
 
-  it('should covert cm2/ms3 to base units', ()=>{
-    const baseUnitData = utils.toBaseUnits('cm2/ms3');
+  it('should convert cm2/ms3 to base units', ()=>{
+    const baseUnitData = utils.convertToBaseUnits('cm2/ms3', 1);
     assert.equal(baseUnitData.msg.length, 0);
     assert.equal(baseUnitData.magnitude, 100000);
+    assert.equal(baseUnitData.fromUnitIsSpecial, false);
     assert.deepEqual(baseUnitData.unitToExp, {'m': 2, 's': -3});
   });
 
 
-  it('should covert B[10.nV] to base units', ()=>{
+  it('should convert B[10.nV] to base units', ()=>{
     // Equivalent to 0.0000316227766016838 g.m2/(C.s2); 1 B[10.nV]=2*log10(10.nV)
-    const baseUnitData = utils.toBaseUnits('B[10.nV]');
+    const baseUnitData = utils.convertToBaseUnits('B[10.nV]', 1);
     assert.equal(baseUnitData.msg.length, 0);
-    assert.equal(baseUnitData.magnitude, undefined);
+    assert.equal(baseUnitData.magnitude.toPrecision(15), 0.0000316227766016838);
+    assert.equal(baseUnitData.fromUnitIsSpecial, true);
     assert.deepEqual(baseUnitData.unitToExp, {'g': 1, 'm': 2, 's': -2, 'C': -1});
   });
 
-  it('should covert [degF] to base units', ()=>{
-    const baseUnitData = utils.toBaseUnits('[degF]');
+  it('should convert [degF] to base units', ()=>{
+    const baseUnitData = utils.convertToBaseUnits('[degF]', 32);
     assert.equal(baseUnitData.msg.length, 0);
-    assert.equal(baseUnitData.magnitude, undefined);
+    assert.equal(baseUnitData.magnitude.toPrecision(15), 273.15);
+    assert.equal(baseUnitData.fromUnitIsSpecial, true);
     assert.deepEqual(baseUnitData.unitToExp, {'K': 1});
   });
 
-  it('should covert 2.[degF] to base units', ()=>{
-    const baseUnitData = utils.toBaseUnits('2.[degF]');
+  it('should convert 2.[degF] to base units', ()=>{
+    const baseUnitData = utils.convertToBaseUnits('2.[degF]', 16);
     assert.equal(baseUnitData.msg.length, 0);
     // Special units do not get a magnitude returned
-    assert.equal(baseUnitData.magnitude, undefined);
+    assert.equal(baseUnitData.magnitude.toPrecision(15), 273.15);
+    assert.equal(baseUnitData.fromUnitIsSpecial, true);
     assert.deepEqual(baseUnitData.unitToExp, {'K': 1});
   });
 
-  it('should covert [degF].2 to base units', ()=>{
-    const baseUnitData = utils.toBaseUnits('[degF].2');
+  it('should convert [degF].2 to base units', ()=>{
+    const baseUnitData = utils.convertToBaseUnits('[degF].2', 16);
     assert.equal(baseUnitData.msg.length, 0);
-    assert.equal(baseUnitData.magnitude, undefined);
+    assert.equal(baseUnitData.magnitude.toPrecision(15), 273.15);
+    assert.equal(baseUnitData.fromUnitIsSpecial, true);
+    assert.deepEqual(baseUnitData.unitToExp, {'K': 1});
+  });
+
+  it('should convert [degF] to base units with a different value', ()=>{
+    const baseUnitData = utils.convertToBaseUnits('[degF]', 41);
+    assert.equal(baseUnitData.msg.length, 0);
+    assert.equal(baseUnitData.magnitude.toPrecision(15), 278.15);
+    assert.equal(baseUnitData.fromUnitIsSpecial, true);
     assert.deepEqual(baseUnitData.unitToExp, {'K': 1});
   });
 
   it('should not covert [iU] to base units', ()=>{
     // [iU] is an "arbitrary" unit and cannot be converted
-    const baseUnitData = utils.toBaseUnits('[iU]');
+    const baseUnitData = utils.convertToBaseUnits('[iU]', 1);
     assert(baseUnitData.msg.length > 0);
     assert.equal(baseUnitData.magnitude, undefined);
+    assert.equal(baseUnitData.fromUnitIsSpecial, undefined);
     assert.deepEqual(baseUnitData.unitToExp, undefined);
   });
 
   it('should not covert m.[iU] to base units', ()=>{
     // [iU] is an "arbitrary" unit and cannot be converted
-    const baseUnitData = utils.toBaseUnits('m.[iU]');
+    const baseUnitData = utils.convertToBaseUnits('m.[iU]', 1);
     assert(baseUnitData.msg.length > 0);
     assert.equal(baseUnitData.magnitude, undefined);
+    assert.equal(baseUnitData.fromUnitIsSpecial, undefined);
     assert.deepEqual(baseUnitData.unitToExp, undefined);
   });
 
   it('should return an error message for a unit string that is a synonym', ()=>{
-    const baseUnitData = utils.toBaseUnits('foot');
+    const baseUnitData = utils.convertToBaseUnits('foot', 1);
     assert(baseUnitData.msg.length > 0);
     assert.equal(baseUnitData.magnitude, undefined);
+    assert.equal(baseUnitData.fromUnitIsSpecial, undefined);
     assert.deepEqual(baseUnitData.unitToExp, undefined);
   });
 
   it('should return an error message for a unit string that is a unrecognized', ()=>{
-    const baseUnitData = utils.toBaseUnits('zzz');
+    const baseUnitData = utils.convertToBaseUnits('zzz', 1);
     assert(baseUnitData.msg.length > 0);
     assert.equal(baseUnitData.magnitude, undefined);
+    assert.equal(baseUnitData.fromUnitIsSpecial, undefined);
     assert.deepEqual(baseUnitData.unitToExp, undefined);
   });
 });